New document for gcc and clang annotation recommendations
Add a document with recommendation for clang and GCC attributes.  This
is a first draft and needs more information to be added to better
describe the implications of each attribute.

The document is currently also missing the standard C attribute names
for many of these attributes, which could potentially make them even
more portable.

Signed-off-by: Siddhesh Poyarekar <siddhesh@gotplt.org>